+++ src/components/LinkChooser/LinkChooser.stories.tsx
@@ -5,8 +5,7 @@ import { Meta, Story } from "@storybook/react";
 import React, { useState } from "react";
 import { LinkChooser as LinkChooserComponent } from "./LinkChooser";
 import { data } from "./mock/data";
-import { guidelineSection } from "./mock/guidelines";
-import { templateSection } from "./mock/templates";
+import { extraSections } from "./sections";
 import { LinkChooserProps, SearchResult } from "./types";
 import { doesContainSubstring } from "./utils/helpers";
 
@@ -36,31 +35,6 @@ const getGlobalByQueryMock = (query: string): Promise<SearchResult[]> =>
             resolve(data.filter((item) => doesContainSubstring(item.title, query, extraSections)));
         }, Math.floor(Math.random() * 2000)),
     );
-
-const getTemplatesByQueryMock = (query: string): Promise<SearchResult[]> =>
-    new Promise((resolve) =>
-        setTimeout(() => {
-            resolve(
-                templateSection.items?.filter((item) => doesContainSubstring(item.title, query, [templateSection])) ||
-                    [],
-            );
-        }, Math.floor(Math.random() * 2000)),
-    );
-
-const getGuidelinesByQueryMock = (query: string): Promise<SearchResult[]> =>
-    new Promise((resolve) =>
-        setTimeout(() => {
-            resolve(
-                guidelineSection.items?.filter((item) => doesContainSubstring(item.title, query, [guidelineSection])) ||
-                    [],
-            );
-        }, Math.floor(Math.random() * 2000)),
-    );
-
-const extraSections = [
-    { ...guidelineSection, getResults: getGuidelinesByQueryMock },
-    { ...templateSection, getResults: getTemplatesByQueryMock },
-];
 
 export const LinkChooser: Story<LinkChooserProps> = (args: LinkChooserProps) => {
     const [openInNewTab, setOpenInNewTab] = useState<boolean>(false);
