+++ src/components/LinkChooser/sections.ts
@@ -1,3 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
+import { guidelineSection } from "./mock/guidelines";
+import { templateSection } from "./mock/templates";
+
 export const defaultSection = { id: "default", title: "Default", items: [] };
+
+export const extraSections = [guidelineSection, templateSection];
